% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFAOSTAT.R
\name{readFAOSTAT}
\alias{readFAOSTAT}
\title{readFAOSTAT}
\usage{
readFAOSTAT(data = "crop", entropy_folder, write = FALSE)
}
\arguments{
\item{data}{Which data to read from faostat}

\item{entropy_folder}{Folder where entropy calculation is made}

\item{write}{}
}
\value{
FAOSTAT cleaned data
}
\description{
readFAOSTAT
}
\examples{
\dontrun{
readFAOSTAT()
}
}
