% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFoodBalances.R
\name{readFoodBalances}
\alias{readFoodBalances}
\title{readFoodBalances}
\usage{
readFoodBalances(src, data = "sua", write = FALSE)
}
\arguments{
\item{src}{Source}
}
\value{
FAOSTAT sua table
}
\description{
readFoodBalances
}
\examples{
\dontrun{
readFoodBalances()
}
}
